+++ libs/traits/Cargo.toml
@@ -16,11 +16,13 @@ codec = { package = 'parity-scale-codec', version = '3.0.0', features = ['derive
 scale-info = { version = "2.0", default-features = false, features = ["derive"] }
 impl-trait-for-tuples = "0.2.1"
 frame-support = { git = "https://github.com/paritytech/substrate", default-features = false , branch = "polkadot-v0.9.29" }
+frame-system = { git = "https://github.com/paritytech/substrate", default-features = false , branch = "polkadot-v0.9.29" }
 sp-runtime = { git = "https://github.com/paritytech/substrate", default-features = false , branch = "polkadot-v0.9.29" }
 sp-arithmetic = { git = "https://github.com/paritytech/substrate", default-features = false , branch = "polkadot-v0.9.29" }
 sp-std = { git = "https://github.com/paritytech/substrate", default-features = false , branch = "polkadot-v0.9.29" }
 orml-asset-registry = {  git = "https://github.com/open-web3-stack/open-runtime-module-library", default-features = false, branch = "polkadot-v0.9.29" }
 cfg-primitives = { path = '../primitives', default-features = false }
+cfg-types = { path = '../types', default-features = false }
 
 # Testing libraries that must be exported
 mockall = { version = "0.11", optional = true }
@@ -28,7 +30,6 @@ lazy_static = { version = "1.4", optional = true }
 
 [dev-dependencies]
 orml-traits = { git = "https://github.com/open-web3-stack/open-runtime-module-library", default-features = true, branch = "polkadot-v0.9.29" }
-cfg-types = { path = '../types', default-features = true }
 
 [features]
 default = ['std']
@@ -41,6 +42,7 @@ std = [
     "sp-arithmetic/std",
     "sp-std/std",
     "cfg-primitives/std",
+    "cfg-types/std",
     "orml-asset-registry/std",
     "mockall",
     "lazy_static",
